+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+### Fixed
+- Fix bad prompt validation for some user inputs; [#219](https://github.com/reliablyhq/cli/issues/219)
 ## [0.12.0] - 2021-04-23
 ### Added
 - Added time observation window for SLO in manifest & prompt user in `slo init` generation; [#1681](https://github.com/reliablyhq/cli/issues/181)